feat(qwen4): qualify native mixed precision and source components

+++ tests/ops/linear/linear_test_common.h
@@ -57,7 +57,7 @@ quantized_weight::PackedWeight make_nvfp4_weight(std::int32_t n, std::int32_t k,
                                                  std::uint32_t seed);
 quantized_weight::PackedWeight make_fp8_weight(std::int32_t n, std::int32_t k, std::uint32_t seed);
 
-void cpu_linear_gemm_fp64(const float* weight, const float* activation, double* output,
+void cpu_linear_gemm_fp64(const double* weight, const float* activation, double* output,
                           std::int32_t n, std::int32_t k, std::int32_t t);
 
 bool cuda_available();
